+++ stats/listeners_test.go
@@ -4,7 +4,9 @@ import (
 	"testing"
 
 	"github.com/blend/go-sdk/assert"
+	"github.com/blend/go-sdk/db"
 	"github.com/blend/go-sdk/logger"
+	"github.com/blend/go-sdk/webutil"
 )
 
 func TestAddWebListeners(t *testing.T) {
@@ -12,9 +14,9 @@ func TestAddWebListeners(t *testing.T) {
 
 	log := logger.None()
 	AddWebListeners(nil, nil)
-	assert.False(log.HasListener(logger.HTTPResponse, ListenerNameStats))
+	assert.False(log.HasListener(webutil.HTTPResponse, ListenerNameStats))
 	AddWebListeners(log, NewMockCollector())
-	assert.True(log.HasListener(logger.HTTPResponse, ListenerNameStats))
+	assert.True(log.HasListener(webutil.HTTPResponse, ListenerNameStats))
 }
 
 func TestAddQueryListeners(t *testing.T) {
@@ -22,9 +24,9 @@ func TestAddQueryListeners(t *testing.T) {
 
 	log := logger.None()
 	AddQueryListeners(nil, nil)
-	assert.False(log.HasListener(logger.Query, ListenerNameStats))
+	assert.False(log.HasListener(db.QueryFlag, ListenerNameStats))
 	AddQueryListeners(log, NewMockCollector())
-	assert.True(log.HasListener(logger.Query, ListenerNameStats))
+	assert.True(log.HasListener(db.QueryFlag, ListenerNameStats))
 }
 
 func TestAddErrorListeners(t *testing.T) {
